+++ main-server/src/http_server/routes/v4/idp.js
@@ -51,7 +51,6 @@ router.post('/callback', validateBody, async (req, res, next) => {
       incoming_request_url,
       incoming_request_status_update_url,
       identity_modification_notification_url,
-      accessor_encrypt_url,
       error_url,
     } = req.body;
 
@@ -59,7 +58,6 @@ router.post('/callback', validateBody, async (req, res, next) => {
       incoming_request_url,
       incoming_request_status_update_url,
       identity_modification_notification_url,
-      accessor_encrypt_url,
       error_url,
     });